+++ desktop_version/src/Logic.h
@@ -1,8 +1,6 @@
 #ifndef LOGIC_H
 #define LOGIC_H
 
-void titleupdatetextcol();
-
 void titlelogic();
 
 void maplogic();
