+++ desktop_version/src/RenderFixed.h
@@ -1,0 +1,14 @@
+#ifndef RENDERFIXED_H
+#define RENDERFIXED_H
+
+void titleupdatetextcol();
+
+void gamerenderfixed();
+
+void titlerenderfixed();
+
+void maprenderfixed();
+
+void gamecompleterenderfixed();
+
+#endif /* RENDERFIXED_H */
